Add menu and conversion logic to currency converter
Introduced a menu display method in ExchangeAPI and refactored Principal to use a switch statement for currency conversion options. Updated Conversor to use double for conversion_rate to support decimal values. These changes improve user interaction and conversion accuracy.

+++ src/com/aluracursos/conversordemonedas/conversor/Conversor.java
@@ -1,4 +1,4 @@
 package com.aluracursos.conversordemonedas.conversor;
 
-public record Conversor(int conversion_rate) {
+public record Conversor(double conversion_rate) {
 }
